requirements.txt
aiogram==2.14.3
Flask==1.1.2
imageai==2.1.6
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
Keras==2.4.3
Markdown==3.3.4
matplotlib==3.3.2
numpy==1.19.3
opencv-python==4.5.1.48
pandas==1.1.5
path==15.1.2
Pillow==7.0.0
pip==21.0.1
python-telegram-bot==13.7
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.11.1
sklearn==0.0
tensorboard==2.4.1
tensorflow-estimator==2.4.0
tqdm==4.60.0
imgaug==0.4.0
